customer controller

+++ src/main/java/com/couponmania2/coupon_project/beans/Company.java
@@ -14,14 +14,14 @@ import java.util.Set;
 public class Company {
 
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(updatable = false)
-    private int id;
+    private long id;
 
     @Column(nullable = false , length = 30)
     private String name;
